Refactor portfolio HTML generation to separate module
- Extracted `generatePortfolioHtml`, `escapeHtml`, and `capitalizeFirstLetter` to `server/lib/portfolio-generator.ts`.
- Updated `server/routes/deploy.ts` to use the extracted function.
- Updated `tests/portfolio-generation.test.ts` to test the real implementation instead of a local copy.
- Adjusted tests to match the actual implementation (capitalization and escaping).
